Web/login register actions (#298)
