+++ bruno/environments/prod.bru
@@ -1,0 +1,4 @@
+vars {
+  base_url: 159.65.125.158
+  port: 8000
+}
