+++ bruno/login.bru
@@ -16,3 +16,7 @@ body:json {
     "password": "pass16"
   }
 }
+
+body:text {
+  {"username":"a","password":"b"}
+}
