+++ bruno/signup.bru
@@ -12,6 +12,7 @@ post {
 
 body:json {
   {
+    "fullName": "asdsa",
     "username": "username16",
     "email": "user16@email.com",
     "password": "pass16"
