+++ web/.env.example
@@ -1,0 +1,2 @@
+VITE_BACKEND_URL=http://localhost
+VITE_PORT=8000
