+++ web/src/index.css
@@ -40,3 +40,9 @@
     letter-spacing: -0.025em;
   }
 }
+
+@layer components {
+  .gradient-mask {
+    mask-image: linear-gradient(0deg, rgba(0, 0, 0, 0), #000);
+  }
+}
