+++ web/src/main.tsx
@@ -8,9 +8,9 @@ import { MantineProvider } from "@mantine/core";
 import "./index.css";
 
 ReactDOM.createRoot(document.getElementById("root") as HTMLElement).render(
-  <React.StrictMode>
-    <MantineProvider classNamesPrefix="znt" theme={theme}>
-      <RouterProvider router={router} />
-    </MantineProvider>
-  </React.StrictMode>
+	<React.StrictMode>
+		<MantineProvider classNamesPrefix="znt" theme={theme}>
+			<RouterProvider router={router} />
+		</MantineProvider>
+	</React.StrictMode>,
 );
