+++ web/src/routes/Home.data.tsx
@@ -1,0 +1,6 @@
+import { makeLoader } from "react-router-typesafe";
+
+export const homeLoader = makeLoader(async ({ request }) => {
+	console.log("homeLoader");
+	return null;
+});
